+++ src/lib/shared_utils/src/canister_specific/individual_user_template/types/hot_or_not/mod.rs
@@ -1,7 +1,7 @@
 use std::{borrow::Cow, cmp::Ordering, collections::BTreeMap, time::SystemTime};
 
 use candid::{CandidType, Decode, Deserialize, Encode, Principal};
-use ic_cdk::api::management_canister::provisional::CanisterId;
+use ic_cdk::api::{call::CallResult, management_canister::provisional::CanisterId};
 use ic_stable_structures::{
     memory_manager::VirtualMemory, storable::Bound, DefaultMemoryImpl, Storable,
 };
@@ -16,10 +16,58 @@ use crate::common::types::{
 };
 
 use super::{
+    arg::PlaceBetArg,
     error::BetOnCurrentlyViewingPostError,
     post::{FeedScore, Post},
-    token::{TokenBalance, TokenTransactions},
+    token::TokenTransactions,
 };
+
+pub trait HotOrNotGame {
+    fn validate_incoming_bet(
+        &self,
+        token: &dyn TokenTransactions,
+        bet_maker_principal: Principal,
+        place_bet_arg: &PlaceBetArg,
+    ) -> Result<(), BetOnCurrentlyViewingPostError>;
+    fn prepare_for_bet(
+        &mut self,
+        bet_maker_principal: Principal,
+        place_bet_arg: &PlaceBetArg,
+        token: &mut dyn TokenTransactions,
+        current_timestamp: SystemTime,
+    ) -> Result<(), BetOnCurrentlyViewingPostError>;
+
+    fn process_place_bet_status(
+        &mut self,
+        bet_response: CallResult<(Result<BettingStatus, BetOnCurrentlyViewingPostError>,)>,
+        place_bet_arg: &PlaceBetArg,
+        token: &mut dyn TokenTransactions,
+        current_timestamp: SystemTime,
+    ) -> Result<BettingStatus, BetOnCurrentlyViewingPostError>;
+
+    fn recieve_bet_from_bet_maker_canister(
+        &mut self,
+        bet_maker_principal_id: Principal,
+        bet_maker_canister_id: Principal,
+        place_bet_arg: &PlaceBetArg,
+        current_timestamp: SystemTime,
+    ) -> Result<BettingStatus, BetOnCurrentlyViewingPostError>;
+
+    fn tabulate_hot_or_not_outcome_for_post_slot(
+        &mut self,
+        post_id: u64,
+        slot_id: u8,
+        token: &mut dyn TokenTransactions,
+        current_timestamp: SystemTime,
+    );
+
+    fn receive_earnings_for_the_bet(
+        &mut self,
+        earnings_amount: u128,
+        hot_or_not_outcome_details: HotOrNotOutcomePayoutEvent,
+        timestamp: SystemTime,
+    );
+}
 
 #[derive(CandidType, PartialEq, Eq, Debug, Serialize, Deserialize)]
 pub enum BettingStatus {
